// engemedical-backend/src/soc/types/CadastroFuncionarioPorSituacao.ts
export interface CadastroFuncionarioPorSituacao {
  CODIGOEMPRESA: string;
  NOMEEMPRESA: string;
  CODIGO: string;
  NOME: string;
  CODIGOUNIDADE: string;
  NOMEUNIDADE: string;
  CODIGOSETOR: string;
  NOMESETOR: string;
  CODIGOCARGO: string;
  NOMECARGO: string;
  CBOCARGO: string;
  CCUSTO: string;
  NOMECENTROCUSTO: string;
  MATRICULAFUNCIONARIO: string;
  CPF: string;
  RG: string;
  UFRG: string;
  ORGAOEMISSORRG: string;
  SITUACAO: string;
  SEXO: string;
  PIS: string;
  CTPS: string;
  SERIECTPS: string;
  ESTADOCIVIL: string;
  TIPOCONTATACAO: string;
  DATA_NASCIMENTO: string;
  DATA_ADMISSAO: string;
  DATA_DEMISSAO: string;
  ENDERECO: string;
  NUMERO_ENDERECO: string;
  BAIRRO: string;
  CIDADE: string;
  UF: string;
  CEP: string;
  TELEFONERESIDENCIAL: string;
  TELEFONECELULAR: string;
  EMAIL: string;
  DEFICIENTE: string;
  DEFICIENCIA: string;
  NM_MAE_FUNCIONARIO: string;
  DATAULTALTERACAO: string;
  MATRICULARH: string;
  COR: string;
  ESCOLARIDADE: string;
  NATURALIDADE: string;
  RAMAL: string;
  REGIMEREVEZAMENTO: string;
  REGIMETRABALHO: string;
  TELCOMERCIAL: string;
  TURNOTRABALHO: string;
  RHUNIDADE: string;
  RHSETOR: string;
  RHCARGO: string;
  RHCCENTROCUSTOUNIDADE: string;
  TPASO?: string;
  DTASO?: string;
  CNPJ?: string;
  NomeLegal?: string;
  DescricaoJornadaSemanal?: string;
  NaturezaAtividade?: string;
}